+++ src/__init__.py
@@ -2,13 +2,15 @@
 
 __version__ = "0.1.0"
 
-from .models.lstm import BidirectionalLSTMModel, LSTMModel
+from .models.lstm import AttentionLSTMModel, BidirectionalLSTMModel, GRUModel, LSTMModel
 from .preprocessing.spline import SplinePreprocessor
 from .training.trainer import Trainer
 
 __all__ = [
     "LSTMModel",
+    "GRUModel",
     "BidirectionalLSTMModel",
+    "AttentionLSTMModel",
     "SplinePreprocessor",
     "Trainer",
 ]